Rename all test files to use first version instead of last version
