+++ functionaltests/8_19_test.go
@@ -21,27 +21,23 @@ import (
 	"testing"
 )
 
-// Data streams get marked for lazy rollover by ES when something
-// changed in the underlying template(s), which in this case is
-// the apm-data plugin update for 8.19 and 9.1:
-// https://github.com/elastic/elasticsearch/pull/119995.
-
-func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
+func TestUpgrade_8_18_to_9_0_Snapshot(t *testing.T) {
 	t.Parallel()
 	from := getLatestSnapshot(t, "8.18")
-	to := getLatestSnapshot(t, "8.19")
+	to := getLatestSnapshot(t, "9.0")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := basicUpgradeLazyRolloverILMTestScenarios(
+	scenarios := basicUpgradeILMTestScenarios(
 		from.Version,
 		to.Version,
 		apmErrorLogs{
 			tlsHandshakeError,
 			esReturnedUnknown503,
 			refreshCache503,
+			// TODO: remove once fixed
 			populateSourcemapFetcher403,
 		},
 	)
@@ -53,22 +49,23 @@ func TestUpgrade_8_18_to_8_19_Snapshot(t *testing.T) {
 	}
 }
 
-func TestUpgrade_8_18_to_8_19_BC(t *testing.T) {
+func TestUpgrade_8_18_to_9_0_BC(t *testing.T) {
 	t.Parallel()
 	from := getLatestVersionOrSkip(t, "8.18")
-	to := getLatestBCOrSkip(t, "8.19")
+	to := getLatestBCOrSkip(t, "9.0")
 	if !from.CanUpgradeTo(to.Version) {
 		t.Skipf("upgrade from %s to %s is not allowed", from.Version, to.Version)
 		return
 	}
 
-	scenarios := basicUpgradeLazyRolloverILMTestScenarios(
+	scenarios := basicUpgradeILMTestScenarios(
 		from.Version,
 		to.Version,
 		apmErrorLogs{
 			tlsHandshakeError,
 			esReturnedUnknown503,
 			refreshCache503,
+			// TODO: remove once fixed
 			populateSourcemapFetcher403,
 		},
 	)
